delete ether file to avoid error of usage
