--- core/src/main/java/org/nervos/appchain/tx/FastRawTransactionManager.java
@@ -1,60 +1,0 @@
-package org.nervos.appchain.tx;
-
-import java.io.IOException;
-import java.math.BigInteger;
-
-import org.nervos.appchain.crypto.Credentials;
-import org.nervos.appchain.protocol.AppChainj;
-import org.nervos.appchain.tx.response.TransactionReceiptProcessor;
-
-/**
- * Simple RawTransactionManager derivative that manages nonces to facilitate multiple transactions
- * per block.
- */
-public class FastRawTransactionManager extends RawTransactionManager {
-
-    private volatile BigInteger nonce = BigInteger.valueOf(-1);
-
-    public FastRawTransactionManager(AppChainj appChainj, Credentials credentials, byte chainId) {
-        super(appChainj, credentials, chainId);
-    }
-
-    public FastRawTransactionManager(AppChainj appChainj, Credentials credentials) {
-        super(appChainj, credentials);
-    }
-
-    public FastRawTransactionManager(
-            AppChainj appChainj, Credentials credentials,
-            TransactionReceiptProcessor transactionReceiptProcessor) {
-        super(appChainj, credentials, ChainId.NONE, transactionReceiptProcessor);
-    }
-
-    public FastRawTransactionManager(
-            AppChainj appChainj, Credentials credentials, byte chainId,
-            TransactionReceiptProcessor transactionReceiptProcessor) {
-        super(appChainj, credentials, chainId, transactionReceiptProcessor);
-    }
-
-    @Override
-    protected synchronized BigInteger getNonce() throws IOException {
-        if (nonce.signum() == -1) {
-            // obtain lock
-            nonce = super.getNonce();
-        } else {
-            nonce = nonce.add(BigInteger.ONE);
-        }
-        return nonce;
-    }
-
-    public BigInteger getCurrentNonce() {
-        return nonce;
-    }
-
-    public synchronized void resetNonce() throws IOException {
-        nonce = super.getNonce();
-    }
-
-    public synchronized void setNonce(BigInteger value) {
-        nonce = value;
-    }
-}
