--- core/src/test/java/org/nervos/appchain/tx/TransferTest.java
@@ -1,66 +1,0 @@
-package org.nervos.appchain.tx;
-
-import java.io.IOException;
-import java.math.BigDecimal;
-
-import org.junit.Before;
-import org.junit.Test;
-
-import org.nervos.appchain.crypto.SampleKeys;
-import org.nervos.appchain.protocol.core.Request;
-import org.nervos.appchain.protocol.core.methods.response.AppGasPrice;
-import org.nervos.appchain.protocol.core.methods.response.TransactionReceipt;
-import org.nervos.appchain.utils.Convert;
-
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
-public class TransferTest extends ManagedTransactionTester {
-
-    private TransactionReceipt transactionReceipt;
-
-    @Before
-    @Override
-    public void setUp() throws Exception {
-        super.setUp();
-        transactionReceipt = prepareTransfer();
-    }
-
-    @Test
-    public void testSendFunds() throws Exception {
-        assertThat(Transfer.sendFunds(appChainj, SampleKeys.CREDENTIALS, ADDRESS,
-                BigDecimal.TEN, Convert.Unit.ETHER).send(),
-                is(transactionReceipt));
-    }
-
-    @Test
-    public void testSendFundsAsync() throws  Exception {
-        assertThat(Transfer.sendFunds(appChainj, SampleKeys.CREDENTIALS, ADDRESS,
-                BigDecimal.TEN, Convert.Unit.ETHER).send(),
-                is(transactionReceipt));
-    }
-
-    @Test(expected = UnsupportedOperationException.class)
-    public void testTransferInvalidValue() throws Exception {
-        Transfer.sendFunds(appChainj, SampleKeys.CREDENTIALS, ADDRESS,
-                new BigDecimal(0.1), Convert.Unit.WEI).send();
-    }
-
-    @SuppressWarnings("unchecked")
-    private TransactionReceipt prepareTransfer() throws IOException {
-        TransactionReceipt transactionReceipt = new TransactionReceipt();
-        transactionReceipt.setTransactionHash(TRANSACTION_HASH);
-        prepareTransaction(transactionReceipt);
-
-        AppGasPrice ethGasPrice = new AppGasPrice();
-        ethGasPrice.setResult("0x1");
-
-        Request<?, AppGasPrice> gasPriceRequest = mock(Request.class);
-        when(gasPriceRequest.send()).thenReturn(ethGasPrice);
-        //when("0x1").thenReturn((Request) gasPriceRequest);
-
-        return transactionReceipt;
-    }
-}
